scheduler/src/requirements.txt
flask==1.0.2
celery==4.2.1
pymongo==3.7.1
jsonschema==2.6.0
PyJWT==1.6.4
cerberus==1.2
yagmail==0.11.214
